+++ NEWS.md
@@ -17,6 +17,7 @@ NEW FEATURES :
 * `removeClusterST()` : remove **New optional time series** + **New additional constraints**    
 * `updateScenarioBuilder()` New type of series "hfl" ("hydro final level", similar to "hydrolevels") is available
 * `removeDistrict()` remove a district from a study
+* `createArea()`/`editArea()`: in API mode, allow the user to customize localization and color of an area
 
 
 NEW FEATURES (other) :